fix(universities): change faculty score fields to float
Update `min_score` and `grant_score` in the `Faculty` model from `BigIntegerField` to `FloatField` to allow for decimal values. Includes a new migration to apply these changes and a seeding script for university data.

+++ universities/models.py
@@ -67,8 +67,8 @@ class Faculty(models.Model):
     )
     name = models.CharField(max_length=255, verbose_name="Nomi")
     description = models.TextField(blank=True, verbose_name="Tavsif")
-    min_score = models.BigIntegerField(default=0, verbose_name="Minimal ball")
-    grant_score = models.BigIntegerField(default=0, verbose_name="Grant ball")
+    min_score = models.FloatField(default=0.0, verbose_name="Minimal ball")
+    grant_score = models.FloatField(default=0.0, verbose_name="Grant ball")
 
     class Meta:
         verbose_name = "Fakultet"
